RePro: Move GET and POST request processing logic to dedicated files
